fix build issues and add dependencies
