+++ app/api/bookings/messages/[id]/route.ts
@@ -1,23 +1,31 @@
+import { NextRequest, NextResponse } from 'next/server'
 import { auth } from '@clerk/nextjs/server'
 import { createClient } from '@/lib/supabase/server'
-import { cookies } from 'next/headers'
 
 export const dynamic = 'force-dynamic'
 
-export async function GET(req: Request, { params }: { params: { id: string } }) {
+export async function GET(
+  req: NextRequest,
+  context: any
+) {
+  const { params } = context as { params: { id: string } }
   const supabase = createClient()
   const { data, error } = await supabase
     .from('booking_messages')
     .select('*')
     .eq('booking_id', params.id)
     .order('timestamp', { ascending: true })
-  if (error) return new Response('Failed', { status: 500 })
-  return Response.json(data)
+  if (error) return NextResponse.json('Failed', { status: 500 })
+  return NextResponse.json(data)
 }
 
-export async function POST(req: Request, { params }: { params: { id: string } }) {
+export async function POST(
+  req: NextRequest,
+  context: any
+) {
+  const { params } = context as { params: { id: string } }
   const { userId } = await auth()
-  if (!userId) return new Response('Unauthorized', { status: 401 })
+  if (!userId) return NextResponse.json('Unauthorized', { status: 401 })
   const { message } = await req.json()
   const supabase = createClient()
   const { error } = await supabase.from('booking_messages').insert({
@@ -25,6 +33,6 @@ export async function POST(req: Request, { params }: { params: { id: string } })
     sender_id: userId,
     message,
   })
-  if (error) return new Response('Failed', { status: 500 })
-  return new Response('ok')
+  if (error) return NextResponse.json('Failed', { status: 500 })
+  return NextResponse.json('ok')
 }
