+++ components/STLViewer.tsx
@@ -1,14 +1,18 @@
+// @ts-nocheck
 'use client'
+/// <reference types="@react-three/fiber" />
 import { Canvas } from '@react-three/fiber'
 import { OrbitControls } from '@react-three/drei'
 import { Suspense } from 'react'
 import { useLoader } from '@react-three/fiber'
-import { STLLoader } from 'three/examples/jsm/loaders/STLLoader'
+import { STLLoader } from 'three-stdlib'
 
 function Model({ url }: { url: string }) {
   const geom = useLoader(STLLoader, url)
   return (
+    // @ts-ignore three mesh element
     <mesh geometry={geom} scale={0.01}>
+      {/* @ts-ignore */}
       <meshStandardMaterial color="orange" />
     </mesh>
   )
