+++ package.json
@@ -11,6 +11,8 @@
   },
   "dependencies": {
     "@clerk/nextjs": "^6.21.0",
+    "@radix-ui/react-dialog": "^1.1.14",
+    "@radix-ui/react-dropdown-menu": "^2.1.15",
     "@react-three/drei": "^9.56.0",
     "@react-three/fiber": "^8.13.7",
     "@supabase/auth-helpers-nextjs": "^0.10.0",
