+++ three-examples.d.ts
@@ -1,0 +1,1 @@
+declare module 'three/examples/jsm/loaders/STLLoader'
