+++ tsconfig.json
@@ -34,6 +34,7 @@
     "**/*.ts",
     "**/*.tsx",
     "next-env.d.ts",
+    "three-examples.d.ts",
     ".next/types/**/*.ts"
   ],
   "exclude": [
